tests: Split live keyboard from module
