+++ tests/unit_tests/pyanaconda_tests/modules/localization/test_module_localization.py
@@ -31,8 +31,6 @@ from pyanaconda.modules.common.structures.requirement import Requirement
 from pyanaconda.modules.common.structures.language import LanguageData, LocaleData
 from pyanaconda.modules.localization.installation import LanguageInstallationTask, \
     KeyboardInstallationTask
-from pyanaconda.modules.localization.live_keyboard import GnomeShellKeyboard, \
-    get_live_keyboard_instance
 from pyanaconda.modules.localization.localization import LocalizationService
 from pyanaconda.modules.localization.localed import LocaledWrapper
 from pyanaconda.modules.localization.localization_interface import LocalizationInterface
@@ -549,137 +547,3 @@ class LocaledWrapperTestCase(unittest.TestCase):
         mocked_system_bus.check_connection.return_value = False
         localed_wrapper = LocaledWrapper()
         self._guarded_localed_wrapper_calls_check(localed_wrapper)
-
-
-class LiveSystemKeyboardTestCase(unittest.TestCase):
-    @patch("pyanaconda.modules.localization.live_keyboard.conf")
-    def test_get_live_keyboard_instance(self, mocked_conf):
-        """Test get_live_keyboard_instance function."""
-        mocked_conf.system.provides_liveuser = True
-        assert isinstance(get_live_keyboard_instance(), GnomeShellKeyboard)
-
-        mocked_conf.reset_mock()
-        mocked_conf.system.provides_liveuser = False
-        assert get_live_keyboard_instance() is None
-
-    def _check_gnome_shell_layouts_conversion(self, mocked_exec_with_capture, system_input, output):
-        mocked_exec_with_capture.reset_mock()
-        mocked_exec_with_capture.return_value = system_input
-
-        gs = GnomeShellKeyboard()
-
-        assert gs.read_keyboard_layouts() == output
-        mocked_exec_with_capture.assert_called_once_with(
-            "gsettings",
-            ["get", "org.gnome.desktop.input-sources", "sources"]
-            )
-
-    def _check_gnome_shell_current_layout_conversion(self, mocked_exec_with_capture, system_input,
-                                                     output):
-        mocked_exec_with_capture.reset_mock()
-        mocked_exec_with_capture.return_value = system_input
-
-        gs = GnomeShellKeyboard()
-
-        assert gs.read_current_keyboard_layout() == output
-        mocked_exec_with_capture.assert_called_once_with(
-            "gsettings",
-            ["get", "org.gnome.desktop.input-sources", "mru-sources"]
-            )
-
-    @patch("pyanaconda.modules.localization.live_keyboard.execWithCaptureAsLiveUser")
-    def test_gnome_shell_keyboard(self, mocked_exec_with_capture):
-        """Test GnomeShellKeyboard live instance layouts."""
-        # test one simple layout set
-        self._check_gnome_shell_layouts_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('xkb', 'cz')]",
-            output=["cz"]
-        )
-
-        # test one complex layout is set
-        self._check_gnome_shell_layouts_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('xkb', 'cz+qwerty')]",
-            output=["cz (qwerty)"]
-        )
-
-        # test multiple layouts are set
-        self._check_gnome_shell_layouts_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('xkb', 'cz+qwerty'), ('xkb', 'us'), ('xkb', 'cz+dvorak-ucw')]",
-            output=["cz (qwerty)", "us", "cz (dvorak-ucw)"]
-        )
-
-        # test layouts with ibus (ibus is ignored)
-        self._check_gnome_shell_layouts_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('xkb', 'cz'), ('ibus', 'libpinyin')]",
-            output=["cz"]
-        )
-
-        # test only ibus layout
-        self._check_gnome_shell_layouts_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('ibus', 'libpinyin')]",
-            output=[]
-        )
-
-        # test wrong input
-        self._check_gnome_shell_layouts_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"wrong input",
-            output=[]
-        )
-
-    @patch("pyanaconda.modules.localization.live_keyboard.execWithCaptureAsLiveUser")
-    def test_gnome_shell_current_keyboard_layout(self, mocked_exec_with_capture):
-        """Test GnomeShellKeyboard live instance current layout."""
-        # test one simple layout is set
-        self._check_gnome_shell_current_layout_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('xkb', 'us')]",
-            output="us"
-        )
-
-        # test one complex layout is set
-        self._check_gnome_shell_current_layout_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('xkb', 'cz+qwerty')]",
-            output="cz (qwerty)"
-        )
-
-        # test multiple layouts are set
-        self._check_gnome_shell_current_layout_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('xkb', 'cz+qwerty'), ('xkb', 'us')]",
-            output="cz (qwerty)"
-        )
-
-        # test layouts with ibus (ibus is ignored)
-        self._check_gnome_shell_current_layout_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('xkb', 'cz'), ('ibus', 'libpinyin')]",
-            output="cz"
-        )
-
-        # test layouts with ibus first (ibus should be skipped)
-        self._check_gnome_shell_current_layout_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('ibus', 'libpinyin'), ('xkb', 'us')]",
-            output="us"
-        )
-
-        # test only ibus layout
-        self._check_gnome_shell_current_layout_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"[('ibus', 'libpinyin')]",
-            output=""
-        )
-
-        # test wrong input
-        self._check_gnome_shell_current_layout_conversion(
-            mocked_exec_with_capture=mocked_exec_with_capture,
-            system_input=r"wrong input",
-            output=""
-        )
